+++ app/src/main/java/com/example/teamtube/Adapter/CategoryVideoAdapter.kt
@@ -6,7 +6,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.example.teamtube.CategoryVideoData.CategoryList
+import com.example.teamtube.Model.CategoryList
 import com.example.teamtube.databinding.CategoryVideoItemBinding
 
 class CategoryVideoAdapter(private val mContext: Context) :
