--- app/src/main/java/com/example/teamtube/CategoryVideoData/CategoryList.kt
@@ -1,37 +1,0 @@
-package com.example.teamtube.CategoryVideoData
-
-/*
-class CategoryList {
-    val categoryList = listOf(
-        "애니메이션",
-        "음악",
-        "동물",
-        "스포츠",
-        "쇼츠",
-        "여행",
-        "게임",
-        "뉴스",
-        "교육",
-        "과학기술",
-        "영화"
-    )
-    val categoryMap = mapOf(
-        "1" to "애니메이션",
-        "10" to "음악",
-        "15" to "동물",
-        "17" to "스포츠",
-        "18" to "쇼츠",
-        "19" to "여행",
-        "20" to "게임",
-        "25" to "뉴스",
-        "27" to "교육",
-        "28" to "과학기술",
-        "30" to "영화"
-    )
-}*/
-
-data class CategoryList(
-    val id : String,
-    val categoryTitle : String,
-    val image : String
-)
