+++ app/src/main/java/com/example/teamtube/Fragment/VideoDetailFragmant.kt
@@ -1,4 +1,4 @@
-package com.example.teamtube
+package com.example.teamtube.Fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
