+++ app/src/main/java/com/example/teamtube/Model/CategoryList.kt
@@ -1,0 +1,7 @@
+package com.example.teamtube.Model
+
+data class CategoryList(
+    val id : String,
+    val categoryTitle : String,
+    val image : String
+)
