+++ app/src/main/java/com/example/teamtube/Model/ChannelModel.kt
@@ -1,4 +1,4 @@
-package com.example.teamtube.ChannelData
+package com.example.teamtube.Model
 
 data class ChannelModel (
     val thumbnails: String,
