--- app/src/main/java/com/example/teamtube/SearchFragmentData/RetrofitClient.kt
@@ -1,22 +1,0 @@
-package com.example.teamtube.SearchFragmentData
-
-import com.google.gson.GsonBuilder
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-object RetrofitClient {
-    private val BASE_URL = "https://www.googleapis.com/youtube/"
-
-    val apiService: INetworkService
-        get() = instance.create(INetworkService::class.java)
-
-    private val instance: Retrofit
-        private get() {
-            val gson = GsonBuilder().setLenient().create()
-
-            return Retrofit.Builder()
-                .baseUrl(BASE_URL)
-                .addConverterFactory(GsonConverterFactory.create(gson))
-                .build()
-        }
-}
