+++ app/src/main/res/layout/fragment_video_detail.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".VideoDetailFragmant">
+    tools:context=".Fragment.VideoDetailFragmant">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/con_layout"
